feat(06-01): top-level page metadata
