+++ src/app/page.tsx
@@ -46,6 +46,17 @@ export const metadata: Metadata = {
     locale: 'en_US',
     type: 'website',
   },
+  twitter: {
+    card: 'summary_large_image',
+    creator: '@hudsor01',
+    title: 'Richard Hudson | Revenue Operations Professional | Dallas-Fort Worth',
+    description:
+      'Richard Hudson: Revenue Operations Professional in Dallas-Fort Worth. Experienced RevOps specialist with $4.8M+ revenue impact and 432% growth delivered. Salesloft Certified Administrator & HubSpot RevOps certified. Expert in sales automation and CRM optimization across 10+ successful projects.',
+    images: ['https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=1200&h=630&fit=crop&crop=face&q=80'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com',
+  },
 }
 
 export default function HomePage() {
