+++ src/app/projects/page.tsx
@@ -1,6 +1,13 @@
 export const dynamic = 'force-static'
 
+import { generateMetadata as genMeta } from '@/app/shared-metadata'
 import { Navbar } from '@/components/layout/navbar'
+
+export const metadata = genMeta(
+  'Projects | Richard Hudson - Revenue Operations Portfolio',
+  'Explore 10+ revenue operations projects by Richard Hudson including pipeline analytics, partner performance dashboards, and CRM optimization case studies.',
+  '/projects'
+)
 import { Footer } from '@/components/layout/footer'
 import { ErrorBoundary } from '@/components/error/error-boundary'
 import { ProjectsPageContent } from './_components/projects-page-content'
