+++ src/app/resume/view/page.tsx
@@ -1,10 +1,11 @@
-import { Metadata } from 'next'
+import { generateMetadata } from '@/app/shared-metadata'
 import ResumeViewClient from './resume-view-client'
 
-export const metadata: Metadata = {
-  title: 'Richard Hudson - Resume PDF',
-  description: 'View and download the complete resume of Richard Hudson, Revenue Operations Professional',
-}
+export const metadata = generateMetadata(
+  'Resume PDF | Richard Hudson',
+  'View and download the complete resume of Richard Hudson, Revenue Operations Professional in Dallas-Fort Worth.',
+  '/resume/view'
+)
 
 export const dynamic = 'force-dynamic'
